User service: Changes in PreAuthorize methods, JwtUtils and JwtAuthFilter added to filter chain

+++ user/src/main/java/com/joaquinrouge/clinics/user/controller/UserController.java
@@ -3,6 +3,7 @@ package com.joaquinrouge.clinics.user.controller;
 import java.util.List;
 import java.util.Set;
 
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -15,27 +16,34 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.auth0.jwt.interfaces.DecodedJWT;
 import com.joaquinrouge.clinics.user.model.Role;
 import com.joaquinrouge.clinics.user.model.User;
 import com.joaquinrouge.clinics.user.service.IUserService;
+import com.joaquinrouge.clinics.user.utils.JwtUtils;
+
+import jakarta.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/api/user")
-@PreAuthorize("hasRole('RECEPTIONIST')")
 public class UserController {
 
     private final IUserService userService;
-
-    public UserController(IUserService userService) {
+    private final JwtUtils jwtUtils;
+    
+    public UserController(IUserService userService,JwtUtils jwtUtils) {
         this.userService = userService;
+        this.jwtUtils = jwtUtils;
     }
 
     @GetMapping
+    @PreAuthorize("hasRole('RECEPTIONIST')")
     public ResponseEntity<List<User>> getAllUsers() {
         return ResponseEntity.ok(userService.findAll());
     }
 
     @GetMapping("/{id}")
+    @PreAuthorize("hasRole('RECEPTIONIST')")
     public ResponseEntity<?> getUserById(@PathVariable Long id) {
         try {
             return ResponseEntity.ok(userService.findById(id));
@@ -45,6 +53,7 @@ public class UserController {
     }
 
     @GetMapping("/username/{username}")
+    @PreAuthorize("permitAll()")
     public ResponseEntity<?> getUserByUsername(@PathVariable String username) {
         try {
             return ResponseEntity.ok(userService.findByUsername(username));
@@ -54,6 +63,7 @@ public class UserController {
     }
 
     @GetMapping("/email/{email}")
+    @PreAuthorize("hasRole('RECEPTIONIST')")
     public ResponseEntity<?> getUserByEmail(@PathVariable String email) {
         try {
             return ResponseEntity.ok(userService.findByEmail(email));
@@ -63,6 +73,7 @@ public class UserController {
     }
 
     @PostMapping("/create")
+    @PreAuthorize("permitAll()")
     public ResponseEntity<?> createUser(@RequestBody User user) {
         try {
             User created = userService.createUser(user);
@@ -73,6 +84,7 @@ public class UserController {
     }
 
     @PutMapping("/add/roles/{id}")
+    @PreAuthorize("hasRole('RECEPTIONIST')")
     public ResponseEntity<?> addRoles(@RequestBody Set<Role> roles,@PathVariable Long id){
     	try {
     		userService.addRoles(id, roles);
@@ -83,6 +95,7 @@ public class UserController {
     }
     
     @PutMapping("/update")
+    @PreAuthorize("hasRole('RECEPTIONIST')")
     public ResponseEntity<?> updateUser(@RequestBody User user) {
         try {
             return ResponseEntity.ok(userService.updateUser(user));
@@ -92,7 +105,24 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<?> deleteUser(@PathVariable Long id,HttpServletRequest request) {
+    	
+    	String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
+		
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(
+            		"Missing or invalid Authorization header");
+        }
+        
+        String token = authHeader.substring(7);
+        DecodedJWT decodedJWT = jwtUtils.validateJWT(token);
+    	
+        if(jwtUtils.getSpecificClaim(decodedJWT, "id").asLong() != id) {
+        	return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(
+            		"Unauthorized");
+        }
+        
         try {
             userService.deleteUser(id);
             return ResponseEntity.noContent().build();
